Add a multidevice partition op that slices off portions of a tensor based on its position along the mesh/cluster axis. (#25570)
### Ticket
#17314

### Problem description
We need an op that is able to slice off portions of the tensor depending
on each device's position along the mesh/cluster_axis. This is needed
for slicing off after broadcasting in DeepSeek, which we need to do as
all_to_all_dispatch and all_to_all_combine assume that real data exists
on every device, not just a single row. To support the two MoE CCLs we
thus need to take the `[1, 1, 64, 7168]` data that exists only on a
single device, and distribute it across the mesh rows, along which there
are 4 devices. We thus need `[1, 1, 16, 7168]` across each device, where
the first quarter is on the first device, second quarter on the second,
and so on.

Secondly, we also need an all gather backwards implementation for the
training team. This op is all gather backwards when the input tensor's
data is the same along the axis.


### What's changed
To support this operation, we will first broadcast all the real data to
every other device via @amorrisonTT 's send/receive and then perform the
partition along the broadcasted cluster axis.

The implementation simply inherits from the slice program config for the
slice op, and deploys it across the different devices with different
begins/ends values. This allows us to use all the existing features in
the slice program config, and even enables us to expand it with more
features if we ever need. As slice adds coverage, so will partition.

### Checklist
- [ ] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes
https://github.com/tenstorrent/tt-metal/actions/runs/16454202296
second run with rename:
https://github.com/tenstorrent/tt-metal/actions/runs/16477459526
- [ ] T3K frequent
https://github.com/tenstorrent/tt-metal/actions/runs/16458634851
second run with rename:
https://github.com/tenstorrent/tt-metal/actions/runs/16477442188
- [ ] New/Existing tests provide coverage for changes