[skip ci] Revert Add  for ttnn pytest splits (#17533) (#25723)
### Ticket
NA

### Problem description
I added a 15MB file to the repo in February with hopes of balancing the
ttnn unit test load across machines.
For a while it worked, and we had better balance, but that balance is
mostly gone now.
We can't afford to keep this file, update it regularly under version
control.

### What's changed
Remove the file, and go back to random split based on number of tests.

### Checklist
- [ ] [ttnn post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/16475100290)
CI passes

Doesn't need to pass, just running it to show the time break down.