Make big-mesh dual-rank test regressions run again (#25794)
### Ticket
Link to Github Issue

### Problem description
It seems with the recent renaming of the Google C++ Test Suites now
cause the tests no longer to be run in regression. 😱

### What's changed
- Update to re-enable the test suites to run in regression.
- This also coincided with recent changes (yesterday) to how custom mesh
graph descriptor was being handled. This also fixes a bug where
`custom_mesh_graph_desc_path_` is populated too late and is now fixed to
populate on `MetalContext` construction.

### Checklist
- [x] All Post Commit:
https://github.com/tenstorrent/tt-metal/actions/runs/16511016457
- [x] T3K Unit Tests:
https://github.com/tenstorrent/tt-metal/actions/runs/16511008778