Support for creating metamodules in the deepseek module framework (#25621)
### !!!!!! Notice for Module writers !!!!!!
This PR introduces a somewhat significant conceptual change to our
deepseek bringup. Please read this description carefully, as it is the
provisional documentation for the new change. Please also migrate your
code to the new API. While this is an annoying change, it is a necessary
one to be able to execute the DeepSeek on a single Galaxy machine.
Hopefully, with the convenience functions, this will be as painless as
possible.

### Background of the issue
Yesterday we noticed that the current architectural design of the
deepseek bringup assumed greater support from the submesh system than
actually exists. Namely, the assumption was that it's possible to
transfer tensors between the submeshes, which is not true.

### The solution (architectural view)
Thus, the new architectural design assumes that decoder blocks in
deepseek will be grouped up into metablocks, calling metalayers
underneath. For example, blocks 3, 18, 33, 48 might have their weights
bunched up together. The idea is that a single activation will be put on
a single row of a galaxy, with the shards on the other devices begin
filled with garbage. The row that the activation will be put on will
directly correspond to the layer that we actually want to call. For
example, if the activation is put on the last row of a galaxy and the
metablock contains blocks 9, 12, 14, 18, the activation will be passed
through the layer 18. This design also allows for later model execution
pipelining.

### The solution (API view)
At the level of a module such as RMSNorm or MLA, all that changes is
that `AbstractModule.convert_weights` now takes in a list of
dictionaries instead of a dictionary. These dictionaries correspond to
the layers that a given metalayer is supposed to handle. For the sake of
testing, the only assumption here is that the number of rows of the
mesh_device is the same as the number of dictionaries given in the
argument.

For the sake of the change being minimally invasive to implement, the
API provides a number of convenience functions in utils.config_helpers
to handle sharding the weights in a form of a list of dictionaries.
Firstly, `get_multidict_key` allows for extracting a key from a list of
dictionaries, to get a list of tensors. Second, a new
`shard_torch_tensor` function provides a more powerful interface for
converting a list of tensors into per-device shards. The resulting
shards are flattened and placed in a row-wise order on the device using
`from_torch_sharded` or `save_sharded_tensor`. I highly recommend
reading the documentation for `shard_torch_tensor`, as it is quite
expansive.

For example, to shard something like an MLP weight, one would use:
```python
weights_list = get_multidict_key(state_dicts, "w1") # List of (hidden_size, intermediate_size)
tensor_shards = shard_torch_tensor(weights_list, (-1, 0), (mesh_device.shape[1], None)) # (32, hidden_size, intermediate_size // 8)
return save_sharded_tensor(dir/"w1.tensor", tensor_shards, ...)
```

### Consequences to the codebase
- The changes to README.md and AbstractModule documentation are missing
for now, will be updated later in the week.
- This change does not break any existing tests, but it is important for
the Module writers to migrate to the new API as this is what the
DecoderBlock module will need to function properly.

### Checklist
- [ ] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes