[Minimal reduce-scatter] Remove hashing on global sems  (#25887)
### Problem description
`reduce_scatter_minimal_async` used to hash on semaphore addresses,
leading to many program cache entries. This made trace capture
difficult, in the case where in trace capture the program is called with
a different global sem than in the compile stage. In addition, the
number of global sems used by this op was unnecessarily tied to the
number of links.

### What's changed
- remove hashing on global sems
- reduce number of global sems used
- update usage in tests & models

### Checklist
- [x] TG demo
https://github.com/tenstorrent/tt-metal/actions/runs/16576533952
- [x] TG frequent
https://github.com/tenstorrent/tt-metal/actions/runs/16569440658
- [x] TG nightly
https://github.com/tenstorrent/tt-metal/actions/runs/16569442636
- [x] T3K nightly
https://github.com/tenstorrent/tt-metal/actions/runs/16569438752
- [x] T3K frequent
https://github.com/tenstorrent/tt-metal/actions/runs/16569445634