Remove MetalContext reintialize in WatcherFixture::TearDown
This reinitialize call races with set_watcher_interval below,
which leads to the new watcher thread polling at full speed.
This in turn increases the possibility of the following race:

1. Watcher polls, reading mailbox.
2. Watcher reads empty launch messages.
3. Device triggers assert.
4. Watcher reads non-empty assertion flag.

This leads to errors such as #25697.

I think for better test isolation it's probably better to always
tear down MetalContext between tests.  The flag to keep errors
seems unnecessary if the MetalContext is only torn down but not
reinitialized, so it is also removed.