Remove MetalContext reintialize in WatcherFixture::TearDown (#25931)
### Ticket
#25697

### Problem description
This `reinitialize` call races with `set_watcher_interval` a few lines
below (which unfortunately sets the interval to 0), causing the new
watcher thread polling at full speed for the *next* test case (yes, the
new watcher thread carries onto the next test). This in turn increases
the probability of the following host-device race:

1. Watcher polls, reading mailbox. This read is not atomic so something
could happen in the middle.
2. Watcher reads empty launch messages, either because no program
started or the first program finished and [launch message was
cleared](https://github.com/tenstorrent/tt-metal/blob/9c105434df65a01ac024db8dfb374569cd244a07/tt_metal/hw/firmware/src/brisc.cc#L524)
and the second program hasn't started.
3. Programs executes quickly on device and triggers an assertion.
4. Watcher reads non-empty assertion flag.

This leads to errors such as
https://github.com/tenstorrent/tt-metal/issues/25697.

### What's changed

The reinitialize is removed so there's no race causing watcher interval
to be 0. Because now we require the first
[RunProgram](https://github.com/tenstorrent/tt-metal/blob/9c105434df65a01ac024db8dfb374569cd244a07/tests/tt_metal/tt_metal/debug_tools/watcher/test_assert.cpp#L157)
call to wait for a watcher dump, this guarantees the next watcher dump
will happen 250ms later (note: because of the race, it wasn't the case
before, and that's why only waiting for a watcher dump did not resolve
the problem). I think it is sufficient for the second program to execute
and trigger the assertion within 250ms.

Instead of reinitialize, the metal context is torn down unconditionally
in `DispatchFixture::TearDown`. I think this is good for better test
isolation, as there shouldn't be any dependencies on MetalContext state
between test cases. This also resulted in a few cleanups:
- keep_errors flag is removed. This is not necessary if MetalContext is
only torn down, but not reinitialized.
- reset_server flag is removed, since the watcher server is always
reset.
- reinitialize in `DPrintFixture::TearDown` is also removed.

Note: Generally, it's still possible to have the host-device race on
mailbox, but it should be unlikely, only when watcher polls very close
to when the program starts. Addressing that probably requires changing
the way host reads mailbox.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/16587698101)
CI passes
- [x] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/16588086057)
CI with demo tests passes (if applicable)