Fix eltwise MUL bcast COLs from calling ZEROACC with no Dest reuse (#546)
### Ticket
[<!-- Link to Github Issue
-->](https://github.com/tenstorrent/tt-metal/issues/25577)

### Problem description
Eltwise MUL bcast COLs was missing an `if` statement for WH which was
causing it to call ZEROACCs which could corrupt the output data in Dest.
-> `if constexpr (binary_reuse_dest !=
EltwiseBinaryReuseDestType::NONE)`
This issue was actually fixed prior for BH, but never back-ported to
WHB0.

### What's changed
Added the omitted check.

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update

### Checklist
<!-- These are required steps and need to be run from tt-metal
repository's Actions. Use links below and replace them with your run -->
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/16604619089)
CI passes