[skip ci] Don't run test-wheels job in package and release (#26042)
### Ticket
NA

### Problem description
package and release workflow is failing due to wheel artifact name
changes. test-wheels workflow does not accept an input for wheel
artifact name.

### What's changed
Since the wheel is thoroughly tested in other ways in package and
release, including in the wheel build stage, there is no need to run
this in package and release.

In the future, we can consider removing test-wheel workflow.

### Checklist
- YOLO