SD3.5 - parallelize CLIP + T5 and integrate into pipeline (#26016)
### Problem description
CLIP and T5 must be parallelized on device in order for weights to fit
in DRAM.

### What's changed
Go TP=4 on CLIP and T5. Reuse a CFG-parallel submesh for CLIP/T5. In the
case where submesh is 2x2, reshape to 1x4. For 4x4 submesh, replicate
CLIP/T5 on rows and use TP=4 on columns.

### Checklist
- [x] T3K Demo
https://github.com/tenstorrent/tt-metal/actions/runs/16626679217

---------

Co-authored-by: DorsaRoh <drohani@tenstorrent.com>