Qwen2.5-VL bringup (#24236)
### Ticket
[#18640](https://github.com/tenstorrent/tt-metal/issues/18640)
#24273 

### Problem description
Bring up Qwen 2.5 VL models for multimodal processing. Target is
`Qwen2.5-VL-72B` on T3K but may bring up smaller models on the way
(including `Qwen2.5-VL-3B` on N150/N300 and `Qwen2.5-VL-32B` on T3K)

Aim is to add multi-image support, no need to add frontend support for
video at this time.

### What's changed
@yieldthought got the code into its first working shape ("0 to 60" as
the saying goes). His commits and some of my early commits (the history
is preserved [here](https://github.com/tenstorrent/tt-metal/pull/21752))
were squashed into
[0e002fc](https://github.com/tenstorrent/tt-metal/pull/24236/commits/0e002fc1d033a91e8fe6549bf4fe5e45f0c9bd54)
to help make the rebasing smoother (all the commits after that one is to
fix issues from the rebasing -- notably, one issue of hung execution and
one issue of garbage output in repeated batchs -- both are tough to
debug). The downside to the smoother rebasing is that the commit history
in this PR does not show Mark's large contribution to Qwen2.5-VL. So, I
want to explicitly shout out to Mark: thank you for getting us to the
first working version of Qwen2.5-VL and thank you for helping me along
the way since then!

There is a big architectural difference from Qwen2.5-VL to Llama
multi-modal models:
- Qwen2.5-VL's vision model processes the image pixels into tokens that
are mixed with the text-prompt tokens; then the mixture of such tokens
are processed by the text model
- Llama multi-modal models are based on cross-attention mechanism (see
this
[article](https://j-qi.medium.com/inside-mllama-3-2-understanding-metas-vision-language-model-architecture-ae12ad24dcbf)
for a detailed explanation)

One interesting challenge that stems from this image-text token mixture
is that the positional embedding of each user's prompts can differ and
must be computed and passed to the text model on a per-user basis.

Both of the above (the arch diff and the positional embedding diff)
required significant changes from the expected use of tt_transformer,
which resulted in separate code for certain modules from
tt_transformers. Though it is tempting to combine the code into
tt_transformer through `if-else` conditionals, I choose to keep them
separate for now until our work towards a clean tt_transformer library
is soon underway.

CI tests are added to cover key demo test cases and key unit tests. 

As part of this PR, the obsolete `models/demos/qwen` is removed.

### What is next

Performance optimization of the model (specially the vision model) is
out of the scope of this PR and is coming in the future in a separate
PR. Progress is tracked by #24274.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes
  - https://github.com/tenstorrent/tt-metal/actions/runs/16652319482
- [x] (For models and ops writers) [Single-card demo
tests](https://github.com/tenstorrent/tt-metal/actions/workflows/single-card-demo-tests.yaml)
CI passes (if applicable) See [recommended dev
flow](https://github.com/tenstorrent/tt-metal/blob/main/models/MODEL_ADD.md#a-recommended-dev-flow-on-github-for-adding-new-models).
- https://github.com/tenstorrent/tt-metal/actions/runs/16652326131 -->
errors are unrelated to this PR
- [x] ~~(bonus) T3K demo tests:
https://github.com/tenstorrent/tt-metal/actions/runs/16658274266~~
- skipping this due to extremely slow T3K CI pipelines; the last time I
ran T3K pipeline, there was a single error with trace region, I fixed
that already.
- [x] New/Existing tests provide coverage for changes
- see:
https://github.com/tenstorrent/tt-metal/pull/24236#issuecomment-3014032730

---------

Co-authored-by: Mark O'Connor <moconnor@tenstorrent.com>