[skip ci] Improve .deb package info (#26151)
### Ticket
N/A

### Problem description
We do not distinguish between different types of builds. That causes
confusion when auditing what's installed/running.
We have conditional logic for ASan/TSan builds relating to stripping
binaries. This logic is incorrect when using a multi-config generator.

### What's changed
* Suffix the version with the build type unless it's Release (the
assumed type).
* Defer the logic that considers the build type for stripping binaries
until later in the process when we are guaranteed to know what type was
built.