add fabric APIs that hide and workaround Blackhole DRAM arbiter bug (#25800)
Issue https://github.com/tenstorrent/tt-metal/issues/25348

There is a dram arbiter bug on Blackhole that can cause noc hang by
dropping and backing up a noc when there are noc reads/writes to the
same DRAM noc endpoint. As a workaround, each noc is assigned different
noc endpoints (cores) to access each given DRAM bank. This bank lookup
is handled automatically for address generators. However, this is a
problem for TT-Fabric.

When TT-Fabric is given a noc address, it doesn't necessarily know if it
is a DRAM address. It must assume the noc coordinate system that was
used for the address. To avoid this bug it needs to detect a DRAM noc
endpoint, and then do a remapping to the new noc endpoint for the noc
used by fabric to perform the DRAM write.

To avoid accidental usability errors, new APIs were added that are more
reflective of the dataflow addrgen APIs. This is slightly innefficient
because an address would be regenerated for each direction on fabric in
the case of a multi-cast. However, upon completion of the API cleanup
and updates
[here](https://github.com/tenstorrent/tt-metal/issues/24850), the number
of address lookups would be reduced back to one.


### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI: https://github.com/tenstorrent/tt-metal/actions/runs/16679805788
- [x] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml)
CI with demo tests passes (if applicable):
https://github.com/tenstorrent/tt-metal/actions/runs/16679802089