Correct End of Page checks in Embedding (#25915)
### Ticket
[Link to Github
Issue](https://github.com/tenstorrent/tt-metal/issues/23481)

### Problem description
Embedding was giving bad pcc for specific shapes, further investigation
found it had to do with shapes with large batch dims for the input
indices

### What's changed
The RM embedding reader was checking only if it was at the end of a
hardware aligned block before checking if it was at the end of a page,
so this only worked for block-aligned input tensors. Now the embedding
kernel correctly checks if we are at the end of a block OR a page.

### Checklist
- [ ] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/16660894070)
CI passes
- [ ] [T3K]()
- [ ] [Device performance
regression](https://github.com/tenstorrent/tt-metal/actions/runs/16579915224)
CI passes (if applicable)