[skip ci] Move fw to metalium pkg (#26166)
### Ticket
A step towards #26144

### Problem description
The fw files were being lumped together with SFPI. However when we get
SFPI from its own package, it won't include those files. The files are
defined in Metalium, so they should ship in a Metalium package.

### What's changed
Shifted the fw files from the JIT package to the Metalium runtime
package.