Fix groupnorm multicore coregrid assert (#26153)
### Problem description
Groupnorm r/c x/y relationship is reversed. This leads to an
inappropriate TT_FATAL assert which limits the coregrid that can be used
by groupnorm.

### What's changed
The initial solution is to fix the relationship between r/c and x/y in
this assert.

Opening PR on behalf of @sadesoyeTT @jonathansuTT 
### Checklist
- [x] APC
https://github.com/tenstorrent/tt-metal/actions/runs/16683430000