add (experimental) lightweight/basic bandwidth telemetry to fabric. (#25638)
This PR adds an opt-in, experimental bandwidth telemetry feature to
TT-Fabric router. The approach here is minimal and future work will
provide more resolution/granularity to the reporting.

The current BW telemetry behaviour is reported once for the entire
workload. It works by timestamping windows where we either send messages
or have unsent/unacknowledged messages in our sender channel. For
windows where no traffic is recorded, the interval is omitted. This is
to avoid reported bandwidth artificially being lowered due to long idle
times in the test.

Results are currently dumped in L1 at the first unreserved address (this
is subject to change) and take the following structure (also subject to
change):

```
struct results {
  uint64_t elapsed_cycles;
  uint64_t zero;
  // eth words are 16B, this lets us have longer capture windows without needing a uin64_t to manage overflow
  uint32_t num_eth_words_sent; 
  uint32_t num_packets_sent;
}
```

Future iterations will add more functionality:
- Finer grain telemetry
- Host polling rate dictates resolution of these bandwidth measurements,
as opposed to one number for a full run
  - Will move it closer to "real-time" reporting from the  
- Per (sender/receiver) channel reporting

## Usage

To enable the telemetry, specifiy environment variable
`TT_METAL_FABRIC_TELEMETRY=1`. This flag will only enable the telemetry
on device side, and will not (currently) enable automatic reporting by
host of telemetry data.

As a rough prototype, the fabric bandwidth tests have implemented a
basic read-back function.

## Limitation
Because this current implementation strictly accumulates into a fixed
set of counters, with no ability to swap out data, there is an integer
overflow concern. Because of this, results are only considered valid if
the test runtime is less than about 10s of runtime. This is fine for our
intended uses cases. This will be lifted in the future as we enable
multiple buffer slots for results that can be round-robin read by host
and cycled through by the router.

## Overheads

Overhead data is still being collected but ranges depending on the
fabric and type of traffic pattern. So far I have observed a 0-1%
overhead. However, further testing may reveal higher max overheads. This
is another reason the feature is opt-in for the time-being.

## Outcomes

Based on the existing telemetry in this change, we are already able to
start plotting useful information for benchmarking and system-level
performance analysis.

For example, this is a line fabric running on the inner 4 devices of a
t3k with traffic being sent in both directions. We see that the middle
hop it "hot" and the end hops are cooler (about half bandwidth), due to
the current fairness policy implemented in the router.

<img width="1309" height="770" alt="image"
src="https://github.com/user-attachments/assets/187387ab-b8c0-4a92-9598-df3ec1ab1e5e"
/>

Here's another example with a full T3K where all rows and columns are
sending data left/right/up/down. Note that the labels/lines may not
exactly line up - the image is really more for illustrative purposes.
Also note that the labels specify "Gbps" but the actual unit of the
numbers are "GB/s". This issue is resolved in this set of changes.
<img width="1394" height="828" alt="image"
src="https://github.com/user-attachments/assets/c2147218-3ad9-40ca-9824-6a0e52f08e2b"
/>

The images above are not intended to represent a productized visualizer
but are more of a quick visualization aid. They are generated by the
`visualize_bw.py` script in this change.


### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI: https://github.com/tenstorrent/tt-metal/actions/runs/16686836591
- [x] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml)
CI with demo tests passes (if applicable):
https://github.com/tenstorrent/tt-metal/actions/runs/16506442679
- [x] Microbenchmark:
https://github.com/tenstorrent/tt-metal/actions/runs/16506444544
- [x] New tests in t3k unit tests:
https://github.com/tenstorrent/tt-metal/actions/runs/16510001431