point-to-point send/receive op (#22880)
### Ticket
https://github.com/tenstorrent/tt-metal/issues/21391
https://github.com/tenstorrent/tt-metal/issues/23425

### Problem description
A point-to-point send/receive Op was required.
 
### What's changed
- send_receive op built.
- A couple of tiny helper functions added to `data_movement/common`
- unit test specifically for MoE use case

### TODO
- fix issues with unaligned RM pages
- ~add semaphore wait to writer to deal with device drift~
- ~Clean up program dispatch so only sender/receiver devices run
kernels~
https://github.com/tenstorrent/tt-metal/issues/25905

Note: it's currently not possible actually allocate a Tensor on a single
device such that it can be used for distributed Ops so at the moment the
inputs and outputs are the single device shards of multi-device mesh
tensors which is awkward but tt_distributed improvements in this area
are forthcoming.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes
https://github.com/tenstorrent/tt-metal/actions/runs/16527136996
- [x] T3K test
https://github.com/tenstorrent/tt-metal/actions/runs/16503868841
- [x] New/Existing tests provide coverage for changes

---------

Co-authored-by: asaigal <asaigal@tenstorrent.com>