SDPA Optimizations: Add mask fusion  (#25916)
### Problem description
This PR does the following optimizations:

We fuses the attention mask addition within the matmul. The output of
the matmul is produced in the DST, we can directly use `acc_to_dst`
feature of `add_tiiles` to directly add the mask on top, hence we do not
need to pack results back to L1 and load it back to DST for adding mask.

### Checklist
- [ ] 4u
[Pipeline](https://github.com/tenstorrent/tt-metal/actions/runs/16701913587)
- [ ] 6u
[Pipeline](https://github.com/tenstorrent/tt-metal/actions/runs/16701963601)