Delete checks for hugepages in tt::Cluster (#26304)
### Ticket

https://github.com/tenstorrent/tt-metal/issues/26194

### Problem description
Hugepages are no longer a requirement for systems with IOMMU enabled.
The check produces false positive on systems with IOMMU but without
enough hugepages.

### What's changed

Deleted the check. I tested on a system with IOMMU disabled & hugepage
disabled, without the check, the program fails at:
https://github.com/tenstorrent/tt-umd/blob/44bd946b34318915b57aadabb509145cde560d6f/device/chip/local_chip.cpp#L163,
so it's won't fail silently.

The check is already a no-op for BH, so deleting it does not affect BH.

### Checklist
- [ ] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/16736215972)
CI passes (ttnn and profiler failures should be unrelated)