[skip ci] Fail fast when the Clang Tidy check fails (#26417)
### Ticket
N/A

### Problem description
Some PRs fail the Clang Tidy check, but the workflow continues on for
what is sometimes many many many extra minutes before it finally reports
a failure for the MQ to evict the PR and re-start all the subsequent
ones. Each minute is wasted resources times the number of queued up
jobs, whose subsequent runs will not only start later, but will be
further behind in the job queue causing even longer delays and making
the MQ pile up even further.

### What's changed
Fail fast.

### Extra thought
It could be limited to only fail fast in the MQ.  Either way is fine.