ci: switch workflow to main branch (#579)
### Ticket
None

### Problem description
With all the fixes now merged into tt-metal, it's time to switch to
using the tt-metal workflow from the main branch.

### What's changed
Implemented workflow cancellation. When a newer workflow is started, the
old one will be cancelled. This time, workflows spanned by the cancelled
workflow will be cancelled as well, so we won't be leaving orphaned runs
that consume CPU.

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update