[skip ci] Shift models-post-commit to CIv2 on WH (#26507)
### Ticket
N/A

### Problem description
N300's are super backlogged in CIv2. We had previously moved TTNN tests
to CIv2, but that backlogged N300's on CIv2 too much that the merge
queue ground to a half, effectively.

### What's changed
Move the Models tests to CIv2. This is a lighter load than the TTNN
tests, so hopefully CIv2 can bear it while still alleviating some
pressure on CIv1.