Simpify JitBuildState and BuildEnvManager
Since most of RISC-specific build options have been moved to hal,
it is redundant to have seperate subclasses of JitBuildState
for DM, Compute, Eth, etc.  If we need different treatment to them,
the logic should be added to hal.

The subclasses are now removed.  As a result, the various constructor
calls in build_env_manager based on core_type and processor_class are
no longer needed, as they are passed as parameter to the constructor
now.

Also fixed an accidental copy of JitBuildState by lambda capture.