Remove legacy unit mesh tests and replace with fabric equivalents (#26809)
### Ticket
#26833

### Problem description
Some level of unit mesh CCLs support is required. We need to support
having tensors allocated on different submeshes and running
all_gather/reduce_scatter on these.

### What's changed
Delete the legacy all gather test as that already has a tested fabric
equivalent in the suite
Replace the Reduce Scatter test with a ReduceScatterAsync test.

### Checklist
- [ ] T3K unit test:
https://github.com/tenstorrent/tt-metal/actions/runs/16949110503