Refactoring to support lite fabric (#26588)
### Ticket
Link to Github Issue

### Problem description
- We would like to re-use the Fabric EDM components for Lite Fabric.
Currently these components will not compile properly because they are
dependent on 1d_fabric_constants and/or the PACKET_HEADER_TYPE which are
not the same on Lite Fabric anymore.

### What's changed
- Template PACKET_HEADER_TYPE in the channel structs
- Template the required constants
- There was a conflict with the DPRINT macro being defined before the
file with the HalL1MemAddrType DPRINT enum got included. Renamed the
enum to DPRINT_BUFFER to fix the build conflict.

### Checklist
APC
https://github.com/tenstorrent/tt-metal/actions/runs/16970221680
BH
https://github.com/tenstorrent/tt-metal/actions/runs/16945815421
T3K
https://github.com/tenstorrent/tt-metal/actions/runs/16864424067
T3K Fast
https://github.com/tenstorrent/tt-metal/actions/runs/16842420315
TG
https://github.com/tenstorrent/tt-metal/actions/runs/16864425385