SD3.5 CI - create empty benchmark result if test is skipped (#26881)
### Problem description
TG model perf pipeline is failing since SD3.5 model perf test is
(purposefully) skipped on 4U
https://github.com/tenstorrent/tt-metal/actions/runs/16966146973/job/48091639762

Higher level, the problem is that the TG and Galaxy pipelines are
coupled. This test was meant to run on 6U only but has to be part of a
4U pipeline, just to get skipped.

### What's changed
On 4U, the test dumps an empty profile. 

CC @subinleeTT 


### Checklist
- [x] TG model perf
https://github.com/tenstorrent/tt-metal/actions/runs/16966146973/job/48091639762