#26763: Correctly upload tensor to device (#26831)
### Ticket
#26763

### Problem description
`to_device()` is not called with correct memory config.

### What's changed
Provide the correct argument and add tests.

Root problem is this subtle behavior:
```cpp
    [[nodiscard]] Tensor to_device(
        distributed::MeshDevice* mesh_device,
        const MemoryConfig& mem_config = MemoryConfig{},
        ttnn::QueueId cq_id = ttnn::DefaultQueueId) const;
```

The default of `mem_config` is default constructed `MemoryConfig` and
not `MemoryConfig` that is present in `Tensor` already. Not clear why.

### Checklist
- [X] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/16954358388)
- [X] New/Existing tests provide coverage for changes