[POOL] Support for arbitrary channel size of the input tensor for pool ops (#26448)
## Tickets
[23581](https://github.com/tenstorrent/tt-metal/issues/23581)
[23617](https://github.com/tenstorrent/tt-metal/issues/23617)

### Problem description
#### How it was working:
Pool operations used to support channel sizes that are either 16 or
multiples of 32. To explain how this was done we need to introduce past
definition of partial tile. Input is having a partial tile if channels
size equals 16. Since the output is always sharded, compute kernel is
handling the job of packing the data in the end. For each output stick
after the reduction compute kernel should pack_untilize from dest
register either one row of 16 datums or one row of 32 datums. It used to
pack_untilize 1 face only if the first tile is partial and two faces in
all the other cases.

### What's changed
#### How it works now:
Instead of partial tile we now have last tile is partial flag where
input tensor can contain one or none partial tile at the end. Input
tensor is considered to have partial tile at the end if in_channels % 32
<= 16.

When received an arbitrary channel size of the input tensor pool op does
the auto sharding, the shard_width is aligned to channels size which is
multiple of 8. This will be the padded shape of the tensor, whereas the
information on the actual channel size is passed all the way to program
factory method. Input shard width will be multiple of 8 where output
shard width will be multiple of 16 due to the fact that pack_untilize
cannot write any less than 1x16 row of data.

Reader kernel is changed to have a different increment throughout in_cb,
in_aligned_nbytes_c which is dependent on width of the page of this cb
but we will read the needed length of the stick so it could be lesser
than page width.

Compute kernel is changed to reserve adequate number of tiles of out_cb
depending on whether the last tile is partial or not. Page size of the
output cb is changed from tile to 16.

#### Models changes:
During this change there were bugs in models exposed, corrected by using
output channel size of the previous op as input channel size for the
following pool op. Moreover an issue when the memory is being corrupted
is detected and fized as well.

### Checklist
- [ ] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/16983512130)
All post commit on same commit main for comparison on the following
[link](https://github.com/tenstorrent/tt-metal/actions/runs/16983400977)
- [x] [(Single-card) Frequent model and ttnn
tests](https://github.com/tenstorrent/tt-metal/actions/runs/16938436801)
- [x] [Device performance
regression](https://github.com/tenstorrent/tt-metal/actions/runs/16938453509)
- [x] [Nightly tt-metal L2
tests](https://github.com/tenstorrent/tt-metal/actions/runs/16938469427)
- [x] [(Blackhole) Blackhole nightly tests
](https://github.com/tenstorrent/tt-metal/actions/runs/16976640343)
- [x] Additional testing which happened locally is measuring duration of
each pool op in resnet model on corresponding commit on main and this
branch which showed that no regressions are introduced.