SDXL: device perf test update (#26999)
### What's changed
Reduced `UNET_DEVICE_TEST_TOTAL_ITERATIONS` in device perf test to 1. 

Overall execution is large enough so that device perf time fluctuations
can be disregarded and SDXL takes a significant portion of the device
perf pipeline; we can decrease this for now and update to larger number
of iterations if there is need for that later.

### Checklist
- [x] [(Single-card) Device perf
regressions](https://github.com/tenstorrent/tt-metal/actions/runs/17035907520)
CI passes