Added support for the grid sample op (#26668)
### Ticket
#21617 

### Problem description
Previously, the grid sample was not supported in the ttnn stack. This
prevented bringup of several models that use this op.

### What's changed
Added support for the grid sample op.

For now, the operation only supports interleaved, row major input and
grid, in bfloat16. Channels are required to be aligned to TILE_WIDTH
(32), and wide reduction is not supported for now (num_channels <= 256
requirement). This still covers all of the shapes requested in the
original issue.

The main perf disadvantage of the current implementation is that the
logic for determining the coords and weights of the input sticks, which
includes floating point arithmetic, is done on the reader RISC. This
poses a significant bottleneck for the operation, which will be fixed in
other PRs.

### Checklist
- [ ] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/16879174680)
CI passes
- [ ] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/16879185647)
CI passes
- [ ] [Device performance
regression](https://github.com/tenstorrent/tt-metal/actions/runs/16879214035)
CI passes
- [ ] [tt-metal L2
tests](https://github.com/tenstorrent/tt-metal/actions/runs/16879198162)
CI passes